smart-vision-demo/requirements.txt
gradio==5.4.0
pymilvus==2.6.2
PyYAML==6.0.2
anytree==2.13.0
imbalanced-learn==0.13.0
lightning==2.5.1
openpyxl==3.1.5
pandas==2.2.3
paddleocr==2.9.1
paddlepaddle-gpu==2.6.1
Pillow==11.3.0
safetensors==0.5.3
scikit-learn==1.6.1
torch==2.7.1
transformers==4.51.3
huggingface-hub==0.34.0
-e ../smart-vision-model
